added transition music
