--- src/components/interstitials/index.ts
@@ -1,1 +1,0 @@
-export * from "./interstitials.jsx";
