+++ src/components/transitions/styles.ts
