+++ src/constants/main.ts
@@ -10,7 +10,8 @@ export const MATCH_TIME_MS = 600000; // 10 minutes
 export const ALERT_TIME_MS = 30000;
 export const VOTING_TIME_MS = 60000;
 
-export const SPLASH_SCREEN_DURATION = 3000; // 3 seconds
+export const SPLASH_SCREEN_DURATION = 6500; // 6.5 seconds
+export const SPLASH_SCREEN_TEXT_DURATION = 500; // 0.5 seconds
 export const MATCHMAKING_DELAY = 1800;
 
 export const DEFAULT_MAX_PLAYERS_PER_ROOM = 5;
